On my sample I see the time fall from 33ms to 16ms. Allocation falls from 500 kB to 380 kB.

* Compute the deduplication by hand instead of using Linq (vast bulk of the improvement)
* Average by dividing rather than normalizing (0.41ms to 0.06ms)
* Don't add the first normal to zero, just assign it (0.16ms to 0.10ms)